PROYECTO RA-2/fase1_extraccion/requirements.txt
# ============================================================
# FASE 1: Extracción de datos
# ============================================================
requests==2.31.0
deltalake==0.19.0
pandas==2.2.0
pyarrow==16.1.0

# ============================================================
# FASE 2: Data Warehouse en NeonDB
# ============================================================
psycopg2-binary==2.9.9  # PostgreSQL adapter
python-dotenv==1.0.0     # Manejo de variables de entorno
tabulate==0.9.0          # Formateo de tablas para validación
